+++ bin/update-conda.sh
@@ -31,16 +31,14 @@ function update_environment {
         # Mac OS
         # Have to replace Mac versions of some programs (date, sed, etc...)
         conda create -y -n bactopia-${1} ${6} -c conda-forge -c bioconda ${2} coreutils sed
-        conda env export --no-builds -n bactopia-${1} | \
-            grep -v "^prefix:" | \
-            sed -E 's=channels:=version: '"${4}"'\'$'\nchannels:=' > ${3}/${1}.yml
+        conda env export --no-builds -n bactopia-${1} > ${3}/${1}.yml
+        echo "Bactopia version: ${4}" > ${3}/${1}.version
         md5 -r ${3}/${1}.yml | cut -d " " -f 1 > ${3}/${1}.md5
     else
         # Linux
         conda create -y -n bactopia-${1} ${6} -c conda-forge -c bioconda ${2} 
-        conda env export --no-builds -n bactopia-${1} | \
-            grep -v "^prefix:" | \
-            sed -r 's=channels:=version: '"${4}"'\nchannels:=' > ${3}/${1}.yml
+        conda env export --no-builds -n bactopia-${1} > ${3}/${1}.yml
+        echo "Bactopia version: ${4}" > ${3}/${1}.version
         md5sum ${3}/${1}.yml | cut -d " " -f 1 > ${3}/${1}.md5
     fi